sqlalchemy for triaging
Load all the triage status data into an in-memory sqlite database with
SQLAlchemy, and generate the triage report with Jinja.

This allows for easier querying and sorting.

In the future, I might make bin/triage use rhcephbugs.triage.models and
persist the data in a sqlite DB on disk. It is really handy to directly
edit the data with Vim, though. Maybe do the TODO stuff in
645de07de72a9606396333659caeebcad5ff6df7 to replace that.